refactored code to respective file
